refactor: DisplayConfigSettings and DisplayApiSettings
