+++ studio/components/layouts/ProjectLayout/BuildingState.tsx
@@ -2,10 +2,6 @@ import Link from 'next/link'
 import { FC, useEffect, useRef } from 'react'
 import { observer } from 'mobx-react-lite'
 import { Badge, IconArrowRight, IconLoader, Button } from '@supabase/ui'
-import {
-  DisplayApiSettings,
-  DisplayConfigSettings,
-} from 'components/to-be-cleaned/DisplayProjectSettings'
 import ExampleProject from 'components/interfaces/Home/ExampleProject'
 import ClientLibrary from 'components/interfaces/Home/ClientLibrary'
 import { CLIENT_LIBRARIES, EXAMPLE_PROJECTS } from 'components/interfaces/Home/Home.constants'
@@ -14,6 +10,7 @@ import { API_URL, PROJECT_STATUS } from 'lib/constants'
 import { useStore } from 'hooks'
 import { get } from 'lib/common/fetch'
 import { Project } from 'types'
+import { DisplayApiSettings, DisplayConfigSettings } from 'components/ui/ProjectSettings'
 
 type ProjectBuildingState = { project: Project }
 const ProjectBuildingState: FC<ProjectBuildingState> = ({ project }) => {
