+++ studio/components/ui/ProjectSettings/SettingsLoadingState.types.tsx
@@ -1,0 +1,4 @@
+export interface SettingsLoadingStateProps {
+  isError: boolean
+  errorMessage: string
+}
