+++ studio/components/ui/ProjectSettings/index.constants.tsx
@@ -1,0 +1,1 @@
+export const DEFAULT_API_SERVICE_ID = 1
