+++ studio/components/ui/ProjectSettings/index.tsx
@@ -1,0 +1,4 @@
+import DisplayApiSettings from './DisplayApiSettings'
+import DisplayConfigSettings from './DisplayConfigSettings'
+
+export { DisplayApiSettings, DisplayConfigSettings }
